pyproject.toml
[project]
name = "vosyanbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Vasiliy Nikiforov",email = "kaktusv6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio (==4.9.0)",
    "certifi (==2025.7.9)",
    "cffi (==1.17.1)",
    "click (==8.2.1)",
    "colorama (==0.4.6)",
    "cryptography (==41.0.3)",
    "debugpy (==1.8.14)",
    "h11 (==0.14.0)",
    "httpcore (==0.17.3)",
    "httpx (==0.24.1)",
    "idna (==3.10)",
    "markdown-it-py (==3.0.0)",
    "mdurl (==0.1.2)",
    "peewee (==3.16.3)",
    "peewee-db-evolve (==3.7.6)",
    "pycparser (==2.22)",
    "pygments (==2.19.2)",
    "pymysql (==1.1.0)",
    "python-dotenv (==1.1.1)",
    "python-telegram-bot (==20.5)",
    "rich (==14.0.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "typer (==0.16.0)",
    "typing-extensions (==4.14.1)",
    "dependency-injector (>=4.48.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
